refactor(webpack): rename webpack.common.js to webpack.config.js
